requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
annotated-types==0.7.0
asyncio==3.4.3
exceptiongroup==1.3.0
idna==3.10
multidict==6.6.3
pamqp==3.3.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
typing-inspection==0.4.1
typing_extensions==4.14.1
yarl==1.20.1
